AnyMapper can use strategies
