+++ src/AnyMapper/AnyMapper.php
@@ -14,7 +14,7 @@ class AnyMapper
     private Proxy $destinationProxy;
     /** @var array<string[]> */
     private array $allowedDataCasting = [];
-    private bool $allowedDynamicProperties = false;
+    private string $strategyClass = StrictStrategy::class;
 
     /**
      * @param mixed[]|object $source
@@ -22,6 +22,16 @@ class AnyMapper
     public function map(array|object $source): self
     {
         $this->sourceProxy = new Proxy($source);
+
+        return $this;
+    }
+
+    /**
+     * @param class-string<StrategyInterface> $strategyClass
+     */
+    public function using(string $strategyClass): self
+    {
+        $this->strategyClass = $strategyClass;
 
         return $this;
     }
@@ -36,9 +46,8 @@ class AnyMapper
     }
 
     /**
-     * @template T of object
-     * @param class-string<T> $className
-     * @return T
+     * @param class-string $className
+     * @return object
      */
     public function as(string $className): object
     {
@@ -83,58 +92,10 @@ class AnyMapper
         return $this;
     }
 
-    public function allowDynamicProperties(bool $allow = true): self
-    {
-        $this->allowedDynamicProperties = $allow;
-
-        return $this;
-    }
-
     private function doMapping(): void
     {
-        $sourceTracks = $this->sourceProxy->getAnalyzer()->getOutputTrackList();
-
-        foreach ($sourceTracks as $trackName => $sourceTrack) {
-            $destinationTrack = $this->destinationProxy->getAnalyzer()->getInputTrack($trackName);
-
-            if (! $destinationTrack && ! $this->allowedDynamicProperties) {
-                continue;
-            }
-
-            $sourceValue = $this->sourceProxy->get($trackName);
-
-            $destinationValue = $this->castOrFail($sourceValue, $destinationTrack);
-
-            $this->destinationProxy->set($trackName, $destinationValue);
-        }
-    }
-
-    private function castOrFail(mixed $value, Track|null $destinationTrack): mixed
-    {
-        $allowedDestinationDataTypes = $destinationTrack?->getDataTypes();
-
-        if (is_null($allowedDestinationDataTypes)) {
-            // Dynamic Properties are allowed, no casting needed
-            assert($this->allowedDynamicProperties);
-            return $value;
-        }
-
-        $sourceType = gettype($value);
-
-        if (in_array($sourceType, $allowedDestinationDataTypes)) {
-            // Source type is already accepted by destination, no casting needed
-            return $value;
-        }
-
-        $allowedDataCasting = $this->allowedDataCasting[$sourceType] ?? [];
-
-        foreach ($allowedDataCasting as $to) {
-            if (! in_array($to, $allowedDestinationDataTypes)) {
-                continue;
-            }
-            return (new Caster($value))->as($to);
-        }
-
-        throw new MapperException('Cannot convert type ' . $sourceType . ' into any of the following types: '. implode(',', $allowedDestinationDataTypes));
+        /** @var StrategyInterface $strategy */
+        $strategy = new $this->strategyClass($this->sourceProxy, $this->destinationProxy, $this->allowedDataCasting);
+        $strategy->run();
     }
 }
