+++ src/AnyMapper/StrategyInterface.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace WebFu\AnyMapper;
+
+interface StrategyInterface
+{
+    public function run(): void;
+}
